brandon-vocal/requirements.txt
anthropic>=0.40.0
elevenlabs>=1.0.0
pvporcupine>=3.0.0
pvrecorder>=1.2.0
rumps>=0.4.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
